+++ catstore-backend/src/main/java/com/catstore/repository/UserRepository.java
@@ -7,6 +7,8 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
+
 @Transactional
 public interface UserRepository extends JpaRepository<User, Integer> {
     @Query(value = "from User where userId = :userId")
@@ -17,9 +19,9 @@ public interface UserRepository extends JpaRepository<User, Integer> {
     void setUserSignature(@Param("userId") Integer userId, @Param("userSignature") String userSignature);
 
     @Query(value = "select userProperty from User where userId =?1")
-    Float getUserPropertyByUserId(Integer userId);
+    BigDecimal getUserPropertyByUserId(Integer userId);
 
     @Modifying
     @Query(value = "update User set userProperty = userProperty + ?2 where userId =?1")
-    void updateUserProperty(Integer userId, Float delta);
+    void updateUserProperty(Integer userId, BigDecimal delta);
 }
